Saved for moment

+++ app/src/main/java/com/alina/clockanimation/utils/Utils.kt
@@ -13,10 +13,10 @@ object Utils {
         return hours * 3600 + minutes * 60 + seconds
     }
 
-    fun formatTime(totalSeconds: Int): String {
-        val hours = TimeUnit.SECONDS.toHours(totalSeconds.toLong())
-        val minutes = TimeUnit.SECONDS.toMinutes(totalSeconds.toLong() - hours * 3600)
-        val seconds = TimeUnit.SECONDS.toSeconds(totalSeconds.toLong() - hours * 3600 - minutes * 60)
+    fun formatTime(totalSeconds: Long): String {
+        val hours = TimeUnit.SECONDS.toHours(totalSeconds)
+        val minutes = TimeUnit.SECONDS.toMinutes(totalSeconds - hours * 3600)
+        val seconds = TimeUnit.SECONDS.toSeconds(totalSeconds - hours * 3600 - minutes * 60)
         return String.format("%02d:%02d:%02d", hours, minutes, seconds)
     }
